+++ node/clients/checkout.ts
@@ -160,9 +160,9 @@ export class Checkout extends JanusClient {
       metric: 'checkout-updateOrderFormCheckin',
     })
 
-  public orderForm = () => {
+  public orderForm = (orderFormId?: string) => {
     return this.post<OrderForm>(
-      this.routes.orderForm(),
+      this.routes.orderForm(orderFormId),
       { expectedOrderFormSections: ['items'] },
       { metric: 'checkout-orderForm' }
     )